tools dnf download and ubuntu22 kver 6.8 build fix
Use curl to get tools and not dnf download to skip checking
package dependancies. Tools (vif) are not meant to be executed within
kernel-build container, but will be copied to host, so we can safely
ignore package dependencies (like glibc version).

Add gcc-12 to build kernels that pins that specific gcc version in the
headers Makefile, like ubuntu22 hwe 6.8 for example.

Change-Id: Ia955f6313bc84d8c83c5026baf299c310cf42b63